Fix integer overflow in sampleOffset accumulation
In GetSampleFileOffset(), accumulated sample sizes were stored in a
uint32_t which wraps on overflow. With attacker-controlled sample sizes
(e.g. 3 samples of 0x80000000 each), summing them wraps to 0, producing
an arbitrary file seek position. Combined with stco/co64 chunk offsets,
this enables out-of-bounds reads from wrong file positions.

Fix:
- Change sampleOffset and m_cachedSfoSampleOffset to uint64_t to prevent
  the accumulation from wrapping.
- Validate the final computed file offset (chunkOffset + sampleOffset)
  against the actual file size before returning.
- Fix Vuln #4 table validation to skip implicit properties when computing
  minimum entry size (was false-positive rejecting valid stsc atoms).

POC: test/poc_offset_overflow.mp4 - video track with 3 samples of size
0x80000000 in one chunk; reading sample 3 accumulates 0x100000000 which
correctly exceeds file size and throws.

+++ src/mp4track.h
@@ -292,7 +292,7 @@ protected:
     // for improved sample file offset query performance
     MP4ChunkId  m_cachedSfoChunkId;
     MP4SampleId m_cachedSfoSampleId;
-    uint32_t    m_cachedSfoSampleOffset;
+    uint64_t    m_cachedSfoSampleOffset;
 
     string m_sdtpLog; // records frame types for H264 samples
 };
